Fix archive page so it works on first run
I noticed that with the code as it was, 11ty would bail out when I
first started the watch process. I was forced to delete the code,
run the watch process, and then reintroduce it. My best theory is that
this has something to do with the limited properties available on each
template object on first run. Possibly one of them throws an error when
you try to access it and this is combined with a bug where the error is
swallowed on first run.

I managed to fix the issue by removing my brute force restructuring of
template objects for use in the template context and instead do a more
targeted restructuring of just the variables I need.